CHROMIUM: DPTF: Add ESIF_LF as an x86 platform driver
This adds the kernel support for the Intel Dynamic Platform
and Thermal Framework (aka Eco-System Independent Framework).

The raw code is imported from the 8.0.1 release from the
following directories:

Products/ESIF_LF/Sources/*
Products/ESIF_LF/Sources/lin/*
Products/ESIF_CM/Sources/*

BUG=chrome-os-partner:25110
TEST=manual: This commit adds new code that is not compiled yet.

When complete this provides a set of kernel drivers that are used
by the ESIF upper framework applications.  This is tested by loading
the drivers and starting the ESIF application to ensure that the
sysfs nodes are created as expected and the DPTF/ESIF application
runs without issue.

Change-Id: I850dcc60374203bfe1706d89b93ccabeb494feaf
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/183042
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Brian Bian <brian.bian@intel.com>