Revert "BACKPORT: drm/i915: update FBC maximum fb sizes"
This reverts commit dcf4377b17d7b8d2ab4ab708aef9c914047287c9.

It seems like the larger FBC sizes don't work at the hw level; they
cause GPU lockups and complete machine freezes. This is fairly
difficult to reproduce, so went unnoticed for a while.

Bug chromium:338588 has been filed to track solving this mystery and
reenabling FBC on link. For now, we just disable it and ship this
thing.

BUG=chromium:328929
TEST=on 8 link machines, run vimeo + webgl aquarium + youtube for 2 hours (so 16 hours of runtime total), no crash

Change-Id: I3098fd620fcb4ea0e4dc88372f7c45a6cf686563
Reviewed-on: https://chromium-review.googlesource.com/184440
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>