CHROMIUM: arm: powergate: Added polling for {un}powergate operation
We have to poll and check the status of the power domains after
toggling the the powergating/unpowergating operation to ensure
that the PMC has done its work. The same check is also needed
when removing the clamp.

This patch solves the problem of pause/play and seek operation while
media file playback on Chrome broswer.

BUG=None
TEST=Tested media file playback on Chrome Browser.

Change-Id: I56061591b180552c719c9640d082414c841e3b91
Signed-off-by: Tushar Khinvasara <tkhinvasara@nvidia.com>
Signed-off-by: Vince Hsu <vinceh@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/184311
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>