DO NOT REBASE: drm/i915/bdw: Extract rp_state_caps logic
We have a need for duplicated parsing of the RP_STATE_CAPS register (and
the setting of the associated fields). To reuse some code, we can
extract the function into a simple helper.

This patch also addresses the fact that we missed doing this for gen8,
something we should have done anyway.

This could be two patches, one to extract, and one to add gen8, but it's
trivial enough that I think one is fine. I will accept a request to
split it. Please notice the fix addressed by v2 below.

Valleyview is left untouched because it is different.

v2: Logically rebased on top of
commit dd0a1aa19bd3d7203e58157b84cea78bbac605ac
Author: Jeff McGee <jeff.mcgee@intel.com>
Date:   Tue Feb 4 11:32:31 2014 -0600

    drm/i915: Restore rps/rc6 on reset

Note with the above change the fix for gen8 is also handled (which was
not the case in Jeff's original patch).

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>

BUG=None
TEST=With entire patch series applied, RPS enables higher framerate on bwtm2

Change-Id: I57990fdc6c90a794220f6b9ed56ff47dcbe302ec
Link: git://people.freedesktop.org/~bwidawsk/drm-intel tag:broadwell-ww11.5
(cherry picked from commit 975c9baf6edd5e0efdf6c97a8b6259c05c1e8aa3)
Reviewed-on: https://chromium-review.googlesource.com/191572
Reviewed-by: James Ausmus <james.ausmus@intel.com>
Tested-by: James Ausmus <james.ausmus@intel.com>
Reviewed-by: Joe Konno <joe.konno@intel.com>
Tested-by: Joe Konno <joe.konno@intel.com>
Reviewed-by: Ben Zhang <benzh@chromium.org>
Commit-Queue: Joe Konno <joe.konno@intel.com>