DO NOT REBASE: drm/i915/bdw: Implement a basic PM interrupt handler
Almost all of it is reusable from the existing code. The primary
difference is we need to do even less in the interrupt handler, since
interrupts are not shared in the same way.

The patch is mostly a copy-paste of the existing snb+ code, with updates
to the relevant parts requiring changes to the interrupt handling. As
such it /should/ be relatively trivial. It's highly likely that I missed
some places where I need a gen8 version of the PM interrupts, but it has
become invisible to me by now.

This patch could probably be split into adding the new functions,
followed by actually handling the interrupts. Since the code is
currently disabled (and broken) I think the patch stands better by
itself.

v2: Move the commit about not touching the ringbuffer interrupt to the
snb_* function where it belongs (Rodrigo)

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>

BUG=None
TEST=With entire patch series applied, RPS enables higher framerate on bwtm2

Change-Id: Id9c13cf3445260542fc118871c5a4e16421f6eaa
Link: git://people.freedesktop.org/~bwidawsk/drm-intel tag:broadwell-ww11.5
(cherry picked from commit 362c665f6282eac8abc9d9447b17c811d2e90f0b)
Reviewed-on: https://chromium-review.googlesource.com/191574
Reviewed-by: James Ausmus <james.ausmus@intel.com>
Tested-by: James Ausmus <james.ausmus@intel.com>
Reviewed-by: Joe Konno <joe.konno@intel.com>
Tested-by: Joe Konno <joe.konno@intel.com>
Reviewed-by: Ben Zhang <benzh@chromium.org>
Commit-Queue: Joe Konno <joe.konno@intel.com>