UPSTREAM: drm: add minor-lookup/release helpers
Instead of accessing drm_minors_idr directly, this adds a small helper to
hide the internals. This will help us later to remove the drm_global_mutex
requirement for minor-lookup.

Furthermore, this also makes sure that minor->dev is always valid and
takes a reference-count to the device as long as the minor is used in an
open-file. This way, "struct file*"->private_data->dev is guaranteed to be
valid (which it has to, as we cannot reset it).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 1616c525b98deb34b8f4b02eccf0ae3a1310fa27)
Signed-off-by: Mark Yao <yzq@rock-chips.com>

BUG=chrome-os-partner:32135, chromium:399935
TEST=Compile test:
   for b in veyron_pinky rush_ryu and samus:
     emerge-${b} chromeos-kernel-3_14
TEST=Sanity check ui on veyron_pinky, rush_ryu and samus

Change-Id: I8d6023966b4f98287d524ba026a17adce040cdb5
Reviewed-on: https://chromium-review.googlesource.com/220163
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>