mfd: tps65090: Stop caching registers
Nearly all of the registers in tps65090 combine control bits and
status bits.  Turn off caching of registers so that we can read status
bits reliably.

NOTE: the IRQnMASK and CG_CTRLn registers are the exception and could
be cached.  If we find that we spend a lot of time reading those we
can turn on cache for just those registers.

BUG=chromium:338657
TEST=Future changes that check tps65090 status now work.

Change-Id: I1850d7bb209a73c2cbe216289c0fe61be2252350
Signed-off-by: Doug Anderson <dianders@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/184606
(cherry picked from commit 55e239ec3f8e4eb8a57aef36ff36155042566f05)
Reviewed-on: https://chromium-review.googlesource.com/184768