CHROMIUM: video: tegra: gk20a: Fix protected elpg race.
Protected elpg calls can be pre-empted in the middle by the deferred
pg_init. This can lead to refcounting errors, which will result in elpg
always remaining on.

Therefore we protect the relevant sections of elpg setup in the deferred
work with a mutex.

BUG=None
TEST=System boots ok

Change-Id: Iaf864785bd6f475cd38a558712fbd40145024ca9
Signed-off-by: Prashant Malani <pmalani@nvidia.com>
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/191868
Tested-by: Vince Hsu <vinceh@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>