CHROMIUM: cros_ec: Expose EC accelerometer data through sysfs
Adding iio driver to expose EC accelerometer up through sysfs.
You can manually read each accelerometer channel, or you can set
up a sysfs trigger to trigger collection of multiple channels at
once and stuff it into a buffer which can be read as a char
device.

Also added to cros_ec a function to read from EC shared memory.

The cros-ec-accel and iio-trig-sysfs modules are automatically
loaded if while register cros-ec we detect that the EC has
accelerometers

BUG=chrome-os-partner:26027
TEST=manual

On chromebook with accelerometer:

$ cd /sys/bus/iio/devices
$ echo 0 > iio_sysfs_trigger/add_trigger
$ cd iio:device0
$ echo "sysfstrig0" > trigger/current_trigger
$ echo 1 > scan_elements/in_angl_en
$ echo 1 > buffer/enable

And trigger a capture with:

$ cd /sys/bus/iio/devices
$ echo 1 > trigger0/trigger_now

Then, read from the char device to make sure we are getting lid angle and timestamp:

$ od -x /dev/iio:device0

Change-Id: I39e61539c3377ed20b95eafa133305cb62b6e48b
Signed-off-by: Alec Berg <alecaberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186307
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>