UPSTREAM: of: add functions to count number of elements in a property
The need to know the number of array elements in a property is
a common pattern. To prevent duplication of open-coded implementations
add a helper static function that also centralises strict sanity
checking and DTB format details, as well as a set of wrapper functions
for u8, u16, u32 and u64.

BUG=None
TEST=build/boot on storm succeded

Suggested-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Heiko Stuebner <heiko.stuebner@bqreaders.com>
Acked-by: Rob Herring <robh@kernel.org>
Acked-by: Grant Likely <grant.likely@linaro.org>
Acked-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Mark Brown <broonie@linaro.org>

Change-Id: I3e355739e64f871168611dee12e5f403d4522332
Reviewed-on: https://chromium-review.googlesource.com/243446
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Tested-by: Narendran Rajan <nrajan@codeaurora.org>
Commit-Queue: Grant Grundler <grundler@chromium.org>