UPSTREAM: MIPS: introduce cpu_coherent_mask
Add a mask of CPUs which are currently known to be operating coherently.
This is setup initially to be all present CPUs, but in a subsequent
patch CPUs in a MIPS Coherent Processing System will be cleared in this
mask as they enter non-coherent idle states. This will be used in order
to determine when a CPU within a CPS system may need to be powered back
up, but may also be used in future to optimise away wakeups for cache
operations or TLB invalidations.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
(cherry picked from commit 76306f4272e036e600254a98bc291df50cedd949)

BUG=None
TEST=Build and boot on Malta

Change-Id: Iae17fdd5dc8270e4abc7a79aa7e5737815d566a3
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/208704
Reviewed-by: Olof Johansson <olofj@chromium.org>