PCI: rcar: Add DMABOUNCE support
Add DMABOUNCE support together with a local memory area for all PCI devices
hanging off this bridge. 4MiB of memory is set aside for coherent
allocations which is shared between the on-chip OHCI and EHCI devices that
are hanging off the PCI bridge. With this patch the driver will work for
all CONFIG_VMSPLIT settings.

Signed-off-by: Magnus Damm <damm@opensource.se>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Acked-by: Simon Horman <horms+renesas@verge.net.au>