UPSTREAM: clk: introduce clk_set_phase function & callback
A common operation for a clock signal generator is to shift the phase of
that signal. This patch introduces a new function to the clk.h API to
dynamically adjust the phase of a clock signal. Additionally this patch
introduces support for the new function in the common clock framework
via the .set_phase call back in struct clk_ops.

BUG=chrome-os-partner:32740
TEST=Build and boot

Change-Id: I11e698cfe30c19969e9cbab0f6ae45314acbe3d3
Signed-off-by: Mike Turquette <mturquette@linaro.org>
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>
Reviewed-by: Heiko Stuebner <heiko@sntech.de>
Acked-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from git.linaro.org/people/mike.turquette/linux.git clk-next
 commit e59c5371fb9d8268d1c043172e88cecab9dc934f)

Reviewed-on: https://chromium-review.googlesource.com/221980
Reviewed-by: Alexandru Stan <amstan@chromium.org>