mmc: dw_mmc: Consider ciu_div when tuning on exynos
If we want to use a ciu_div less than 3 (divide by 4) we need to
adjust our tuning algorithm.  If we've got a ciu_div of 2 (divide by
3) we only can use 6 samples.  If we've got a ciu_div of 1 (divide by
2) we only have 4 samples.  A ciu_div of 0 (divide by 1) should have 2
samples but is documented to only have 1.

If we've got a ciu_div of > 3 we can still choose 8 samples but we
can't produce later phase offsets.  We'll assume that the existing
algorithms will work OK in that case.

BUG=chrome-os-partner:25782, chrome-os-partner:25364
TEST=See Icfbd801304a704e9cb7f0048330e471d4faf485d

Change-Id: I2e70ed094b771ff9018101b4fa93629626fb95e9
Signed-off-by: Doug Anderson <dianders@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/187319
(cherry picked from commit 2ffb36d549e0a58797d638a45e03e029c17f424a)
Reviewed-on: https://chromium-review.googlesource.com/187959