CHERRY-PICK: CHROMIUM: Fix IIO buffer bug in which test_bit() is used to get scan mask.
Fixes bug in iio buffer. In a couple places, test_bit() is used to
check the scan mask and is expected to return 0 or 1 due to the sysfs
API. But, test_bit() actually returns simply true or false. On some
platforms, true will be negative, which is not expected and causes
problems. This fix ensures that the scan mask is read as only 0 or 1.

BUG=none
TEST=manual test on glimmer. test_bit() returns -1 or 0 on glimmer, so
this fix makes it such that when you read the scan mask it returns 0 or
1 instead of 0 or -1.

Original-Change-Id: I8cf0bdfb8decd82439c8f9454c8c22ce08e64113
Signed-off-by: Alec Berg <alecaberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187673
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
(cherry picked from commit d284bdac065d6b5d1349a5fe00789d6eea62e093)

Change-Id: I33d2b5a4910882dbdf415f7b77c1ad533ccef3f9
Reviewed-on: https://chromium-review.googlesource.com/189396
Commit-Queue: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Heng-ruey Hsu <henryhsu@chromium.org>