CHROMIUM: pstore: Move to larger record size (again)
Accidentally specified a smaller record size than before, bring it back
to the same size as we had when we used the config file.

This was already done in CL 170128, but the switch over to hardcoded values
was based on earlier code so the change was lost in the move over.

BUG=chromium:295127
TEST=none

Change-Id: I3719c2a4d4cd8ff3f18c4a2ee53f9aecbf8058d2
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188075
Reviewed-by: Sameer Nanda <snanda@chromium.org>
[ bleung : pick to R34 branch for chromeos-3.10 ]
(cherry picked from commit 69c6831bf3705d7ff301ee7cb256733a98d099e1)
Reviewed-on: https://chromium-review.googlesource.com/188461
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>