UPSTREAM: ASoC: dapm: Consolidate MUXs and virtual MUXs
MUXs and virtual MUXs are almost identical, the only difference is that for
virtual MUX there is no hardware backing register in which setting is stored.
This patch adds code, which is similar to what we already do for DAPM mixer
controls to support virtual mixer controls, to DAPM enum controls. The new code
will check if the enum does a hardware backing register and skip over reading
and writing to the register if it has not.  This allows us to use the same code
path for both MUXs and virtual MUXs and a lot of nearly identical code can be
removed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit 236aaa6863581634bd6d599ccf7f7b38deeafdc0)
Signed-off-by: Kevin Strasser <kevin.strasser@intel.com>

BUG=chrome-os-partner:29649
TEST=With entire patch series applied, audio playback is working

Change-Id: I6c78855398970d7fdefccb228d2dca5a07de549f
Reviewed-on: https://chromium-review.googlesource.com/215177
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Ben Zhang <benzh@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>