drm: Support EDID quirks for reducing blanking
On rambi, the panel timings are such that we are just barely above
what a single 1.6G lane can do, so we need to run at 2.7G. This patch
adds the ability to have panel quirks to reduce the panel blanking.

We also add a quirk to reduce the rambi panel timings and make it fit
in a single 1.6G lane. This saves ~70mW (ballpark, it's very difficult
to get good rambi power measurements).

BUG=none
TEST=instrumented the code and checked that the timing was reduced
TEST=measured power with power_supply_info

Change-Id: Idb3e3da71f9cb11347d886e2b0ee5dd33f9748bf
Reviewed-on: https://chromium-review.googlesource.com/188448
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>