mm: close PageTail race
Since "mm, compaction: avoid isolating pinned pages", it has been possible
for page_count(page) to race with prep_compound_page() by finding
PageTail(page) set with a NULL or dangling page->first_page.

"mm, page_alloc: make first_page visible before PageTail" adds a store
memory barrier to prep_compound_page() to ensure page->first_page is set,
but nothing is preventing compound_head() from seeing a dangling head
page.

This patch uses Andrea's implementation of compound_trans_head() that
deals with such a race and makes it the default compound_head()
implementation.  This includes a read memory barrier that ensures that if
PageTail(head) is true that we return a head page that is neither NULL nor
dangling.

This is the safest way to ensure we see the head page that we are
expecting, PageTail(page) is already in the unlikely() path and the memory
barriers are unfortunately required.

Hugetlbfs is the exception, we don't enforce a store memory barrier during
init since no race is possible.

Signed-off-by: David Rientjes <rientjes@google.com>
Cc: Holger Kiehl <Holger.Kiehl@dwd.de>
Cc: Christoph Lameter <cl@linux.com>
Cc: Rafael Aquini <aquini@redhat.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Michal Hocko <mhocko@suse.cz>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Rik van Riel <riel@redhat.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>